Revert "(refactor) Refactor duplicate UI Authen files into a shared component…"
