--- AGDPMS.Shared/Components/ForgotPassword.razor
@@ -1,78 +1,0 @@
-@page "/forgot-password"
-@using AGDPMS.Shared.Services
-@inject IApiClient ApiClient
-@inject NavigationManager Navigation
-
-<div class="forgot-password-page">
-    <div class="forgot-password-container">
-        <h1>Forgot your password?</h1>
-        <p>Enter your phone number.</p>
-        
-        <div class="form-group">
-            <label for="phoneNumber">Phone Number</label>
-            <input type="tel" id="phoneNumber" @bind="phoneNumber" class="form-control" />
-        </div>
-        
-        @if (!string.IsNullOrWhiteSpace(message))
-        {
-            <div class="error-message">@message</div>
-        }
-        
-        <button @onclick="SendSms" class="btn btn-primary" disabled="@isLoading">
-            @if (isLoading)
-            {
-                <span>Sending SMS...</span>
-            }
-            else
-            {
-                <span>Send SMS</span>
-            }
-        </button>
-        
-        <button @onclick="BackToLogin" class="btn btn-secondary">Back to Login</button>
-    </div>
-</div>
-
-@code {
-    private string phoneNumber = string.Empty;
-    private string? message;
-    private bool isLoading = false;
-
-    private async Task SendSms()
-    {
-        message = null;
-        isLoading = true;
-        
-        try
-        {
-            var result = await ApiClient.ForgotPasswordAsync(new ForgotPasswordRequest
-            {
-                Phone = phoneNumber
-            });
-
-            if (result?.Success == true && result.UserId.HasValue)
-            {
-                Navigation.NavigateTo($"/verify-otp/{result.UserId.Value}");
-            }
-            else
-            {
-                message = result?.Message ?? "Failed to send SMS";
-            }
-        }
-        catch (Exception ex)
-        {
-            message = $"API Error: {ex.Message}";
-        }
-        finally
-        {
-            isLoading = false;
-        }
-    }
-
-    private void BackToLogin()
-    {
-        Navigation.NavigateTo("/login");
-    }
-}
-
-
