--- AGDPMS.Shared/Components/UserProfile.razor
@@ -1,47 +1,0 @@
-@page "/profile"
-@using AGDPMS.Shared.Services
-@inject IUserSession UserSession
-@inject NavigationManager Navigation
-@inject IFormFactor FormFactor
-
-<h2>User Profile</h2>
-
-@if (_profile is null)
-{
-    <p>Not logged in.</p>
-}
-else
-{
-    <ul>
-        <li>User ID: @_profile.UserId</li>
-        <li>Full Name: @_profile.FullName</li>
-        <li>Phone: @_profile.PhoneNumber</li>
-        <li>Role: @_profile.RoleName</li>
-    </ul>
-
-    <button class="btn btn-primary" @onclick="ChangePassword">Change Password</button>
-}
-
-@code {
-    private AGDPMS.Shared.Services.UserProfileDto? _profile;
-
-    protected override async Task OnInitializedAsync()
-    {
-        _profile = await UserSession.GetCurrentAsync();
-    }
-
-    private void ChangePassword()
-    {
-        var form = FormFactor.GetFormFactor();
-        if (string.Equals(form, "Web", StringComparison.OrdinalIgnoreCase))
-        {
-            Navigation.NavigateTo("/need-change-password");
-            return;
-        }
-
-        var id = _profile?.UserId ?? 0;
-        Navigation.NavigateTo($"/need-change-password/{id}");
-    }
-}
-
-
