+++ AGDPMS.Shared/Services/AuthDtos.cs
@@ -67,12 +67,6 @@ public sealed class ChangePasswordRequest
     public string NewPassword { get; set; } = string.Empty;
 }
 
-public sealed class ChangePasswordResponse
-{
-    public bool Success { get; set; }
-    public string? Message { get; set; }
-}
-
 public sealed class AddAccountRequest
 {
     public string PhoneNumber { get; set; } = string.Empty;
