+++ AGDPMS.Web/Components/Account/Pages/ForgotPassword.razor
@@ -1,0 +1,65 @@
+﻿@page "/forgot-password"
+
+@using System.ComponentModel.DataAnnotations
+@using Microsoft.Extensions.Caching.Memory
+@using AGDPMS.Web.Data
+@using AGDPMS.Shared.Services
+
+@inject UserDataAccess UserDataAccess
+@inject ISmsSender SmsSender
+@inject IMemoryCache Cache
+@inject NavigationManager Nav;
+
+<PageTitle>Forgot your password?</PageTitle>
+
+<h2>Forgot your password?</h2>
+<h3>Enter your phone number.</h3>
+
+@if (!string.IsNullOrEmpty(Message))
+{
+    <p>@Message</p>
+}
+<hr />
+
+<EditForm Model="Input" method="post" OnValidSubmit="HandleForgotPassword" FormName="forgot-password">
+    <DataAnnotationsValidator />
+    <ValidationSummary />
+    <div>
+        <label for="Input.Phone">Phone Number</label>
+        <InputText @bind-Value="Input.Phone" id="Input.Phone" />
+    </div>
+    <a href="/login">Cancel</a>
+    <button type="submit">Send SMS</button>
+</EditForm>
+
+@code {
+    [SupplyParameterFromForm]
+    private InputModel Input { get; set; } = new();
+
+    private string? Message;
+
+    private async Task HandleForgotPassword()
+    {
+        var user = await UserDataAccess.GetByPhoneNumberAsync(Input.Phone);
+        if (user is null)
+            return;
+        try
+        {
+            var otp = Random.Shared.Next(100000, 999999).ToString();
+            Console.WriteLine(otp);
+            Cache.Set($"otp:{user.Id}", otp, TimeSpan.FromMinutes(5));
+            await SmsSender.SendAsync($"Your reset code is: {otp}", [user.PhoneNumber]);
+        } catch (Exception e)
+        {
+            Message = e.Message;
+            return;
+        }
+        Nav.NavigateTo($"/verify-otp/{user.Id}");
+    }
+
+    private sealed class InputModel
+    {
+        [Required, Phone]
+        public string Phone { get; set; } = string.Empty;
+    }
+}
