+++ AGDPMS.Web/Components/Account/Pages/NeedChangePassword.razor
@@ -3,7 +3,6 @@
 @using Microsoft.AspNetCore.Authentication
 @using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Identity
-@using Microsoft.Extensions.Caching.Memory
 @using System.ComponentModel.DataAnnotations
 @using System.Security.Claims
 @using AGDPMS.Web.Data
@@ -13,7 +12,6 @@
 @inject UserDataAccess UserDataAccess
 @inject IPasswordHasher<AppUser> PasswordHasher
 @inject NavigationManager Nav
-@inject IMemoryCache Cache
 
 <h3>NeedChangePassword</h3>
 
@@ -66,10 +64,16 @@
         var hash = PasswordHasher.HashPassword(user, Input.NewPassword);
         await UserDataAccess.SetPasswordHashAsync(user.Id, hash, needChange: false);
 
-        var token = Guid.NewGuid().ToString("N");
-        Cache.Set($"signin:{token}", user.Id, TimeSpan.FromMinutes(5));
-        var callbackUrl = $"/api/auth/callback?token={token}&returnUrl={Uri.EscapeDataString("/")}";
-        Nav.NavigateTo(callbackUrl, forceLoad: true);
+        var claims = user.ToClaims();
+        var identity = new ClaimsIdentity(claims, Constants.AuthScheme);
+        var principal = new ClaimsPrincipal(identity);
+        var authProps = new AuthenticationProperties
+        {
+            IsPersistent = true
+        };
+
+        await HttpContext.SignInAsync(Constants.AuthScheme, principal, authProps);
+        Nav.NavigateTo("/");
     }
 
     private class InputModel
