+++ AGDPMS.Web/Components/Pages/Home.razor
@@ -27,7 +27,5 @@
 Welcome to your new app.
 
 <a href="/accounts">Accounts</a>
-|
-<a href="/profile">Profile</a>
 <br />
 <button type="button" @onclick='() => Nav.NavigateTo("/logout", forceLoad: true)'>Logout</button>
