+++ AGDPMS/Components/App.razor
@@ -3,9 +3,8 @@
 @using AGDPMS.Components.Layout
 @using AGDPMS.Components.Pages
 @inject NavigationManager Navigation
-@inject AGDPMS.Services.NavigationHistoryService NavHistory
 
-<Router AppAssembly="@typeof(App).Assembly" AdditionalAssemblies="new[] { typeof(AGDPMS.Shared.Routes).Assembly }">
+<Router AppAssembly="@typeof(App).Assembly">
     <Found Context="routeData">
         <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
     </Found>
@@ -18,38 +17,13 @@
 </Router>
 
 @code {
-    private bool _didRedirect;
-
-    protected override async Task OnAfterRenderAsync(bool firstRender)
+    protected override void OnInitialized()
     {
-        if (firstRender)
+        // Check if user is remembered and redirect accordingly
+        var remembered = Microsoft.Maui.Storage.Preferences.Default.Get("remembered", false);
+        if (!remembered && Navigation.Uri.EndsWith("/"))
         {
-            // Prefer SecureStorage; fall back to Preferences for compatibility
-            bool isRemembered = false;
-            try
-            {
-                var remembered = await Microsoft.Maui.Storage.SecureStorage.Default.GetAsync("remembered");
-                isRemembered = string.Equals(remembered, "true", StringComparison.OrdinalIgnoreCase);
-            }
-            catch { /* ignore */ }
-
-            if (!isRemembered)
-            {
-                var prefRemembered = Microsoft.Maui.Storage.Preferences.Default.Get("remembered", false);
-                isRemembered = prefRemembered;
-            }
-
-            if (!isRemembered && Navigation.Uri.EndsWith("/") && !_didRedirect)
-            {
-                _didRedirect = true;
-                Navigation.NavigateTo("/login", replace: true);
-            }
-
-            NavHistory.Attach(Navigation);
-            Navigation.LocationChanged += (_, args) =>
-            {
-                NavHistory.Push(args.Location);
-            };
+            Navigation.NavigateTo("/login");
         }
     }
 }
